Refactor: promote Injectables out of Cli package
