+++ src/bin/storyplayer
@@ -20,11 +20,12 @@ use Phix_Project\CliEngine\Switches\ShortHelpSwitch;
 use Phix_Project\CliEngine\Switches\VersionSwitch;
 use Phix_Project\CliEngine\Commands\HelpCommand;
 
+use DataSift\Storyplayer\Injectables;
+
 use DataSift\Storyplayer\Cli\Automate_Command;
 use DataSift\Storyplayer\Cli\CreateStory_Command;
 use DataSift\Storyplayer\Cli\DeviceListHelper;
 use DataSift\Storyplayer\Cli\ExtraConfigHelper;
-use DataSift\Storyplayer\Cli\Injectables;
 use DataSift\Storyplayer\Cli\Install_Command;
 use DataSift\Storyplayer\Cli\KillProcesses_Switch;
 use DataSift\Storyplayer\Cli\ListHostsTable_Command;
