+++ src/php/DataSift/Storyplayer/Injectables/ActiveLocalEnvironmentConfigSupport.php
@@ -34,73 +34,55 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   Storyplayer/Cli
+ * @package   Storyplayer/Injectables
  * @author    Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright 2011-present Mediasift Ltd www.datasift.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
 
-namespace DataSift\Storyplayer\Cli;
+namespace DataSift\Storyplayer\Injectables;
 
 use DataSift\Stone\ObjectLib\BaseObject;
 
 /**
- * support for working with the list of known systems-under-test
+ * support for the local environment that the user chooses
  *
  * @category  Libraries
- * @package   Storyplayer/Cli
+ * @package   Storyplayer/Injectables
  * @author    Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright 2011-present Mediasift Ltd www.datasift.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
-trait Injectables_KnownSystemsUnderTestSupport
+trait ActiveLocalEnvironmentConfigSupport
 {
-	public $knownSystemsUnderTest;
-	public $knownSystemsUnderTestList = array();
+	public $activeLocalEnvironmentName;
+    public $activeLocalEnvironmentConfig;
 
-	public function initKnownSystemsUnderTestSupport($additionalSuts)
+	public function initActiveLocalEnvironmentConfigSupport($envName, $injectables)
 	{
-		// special case:
-		//
-		// did we find any systems-under-test in the config files?
-		//
-		// if we didn't, then we create a fake system-under-test entry
-		// so that Storyplayer will run. this is better for new users
-		// than Storyplayer throwing an error because they haven't created
-		// any systems-under-test in the configs
-		if (count($additionalSuts) == 0) {
-			$this->knownSystemsUnderTest = new KnownSystemsUnderTest;
+        // does the local environment exist?
+        if (!isset($injectables->knownLocalEnvironments->$envName)) {
+            throw new E4xx_NoSuchLocalEnvironment($envName);
+        }
 
-			foreach ($this->knownSystemsUnderTest as $name => $config) {
-				$this->knownSystemsUnderTestList[$name] = $name;
-			}
+        // a helper to load the config
+        $staticConfigManager = $injectables->staticConfigManager;
 
-			// all done
-			return;
-		}
+        // build the environment that we want
+        if (is_string($injectables->knownLocalEnvironments->$envName)) {
+            $this->activeLocalEnvironmentConfig = $staticConfigManager->loadConfigFile(
+                $injectables->knownLocalEnvironments->$envName
+            );
+        }
+        else {
+            $this->activeLocalEnvironmentConfig = $injectables->knownLocalEnvironments->$envName;
+        }
 
-		// if we get here, then we ONLY want the systems-under-test that
-		// have been explicitly defined in the configs
-		//
-		// the hardcoded systems-under-test in the KnownSystemsUnderTest
-		// class are only a fallback for when there is NO config at all
-		//
-		// we start with NO known systems under test
-		$this->knownSystemsUnderTest = new BaseObject;
+        // remember the environment name, just in case
+        $this->activeLocalEnvironmentName = $envName;
 
-		// now add in all the systems-under-test that we have discovered
-		// in the config files
-		foreach ($additionalSuts as $filename => $config) {
-			$sutName = basename($filename, 'json');
-			$this->knownSystemsUnderTestList[$sutName] = $sutName;
-			$this->knownSystemsUnderTest->$sutName     = $config;
-		}
-
-		// now put the list of systems-under-test into a sensible order
-		ksort($this->knownSystemsUnderTestList);
-
-		// all done
+        // all done
 	}
 }
