+++ src/php/DataSift/Storyplayer/Injectables/DefaultConfigSupport.php
@@ -34,47 +34,38 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   Storyplayer/Cli
+ * @package   Storyplayer/Injectables
  * @author    Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright 2011-present Mediasift Ltd www.datasift.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
 
-namespace DataSift\Storyplayer\Cli;
+namespace DataSift\Storyplayer\Injectables;
+
+use DataSift\Storyplayer\Cli\DefaultConfig;
 
 /**
- * determine our default target environment to test against
+ * support for our built-in defaults
  *
  * @category  Libraries
- * @package   Storyplayer/Cli
+ * @package   Storyplayer/Injectables
  * @author    Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright 2011-present Mediasift Ltd www.datasift.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
-trait Injectables_DefaultTestEnvironmentName
+trait DefaultConfigSupport
 {
-	public $defaultTestEnvironmentName;
+	public $defaultConfig;
 
-	public function initDefaultTestEnvironmentName($injectables)
+	public function initDefaultConfigSupport()
 	{
-		// what are the candidates?
-		$searchList = [
-			HostnameHelper::getHostname(),
-			"localhost"
-		];
+		// create our default config - the config that we'll use
+		// unless the config file on disk overrides it
+		$this->defaultConfig = new DefaultConfig();
 
-		// do any of these environments exist?
-		foreach ($searchList as $testEnvName) {
-			if (isset($injectables->knownTestEnvironments->$testEnvName)) {
-				$this->defaultTestEnvironmentName = $testEnvName;
-				// all done
-				return;
-			}
-		}
-
-		// if we get here, then none of the environments exist
-		$this->defaultTestEnvironmentName = end($searchList);
+		// all done
+		return $this->defaultConfig;
 	}
 }
