+++ src/php/DataSift/Storyplayer/Injectables/HostnameHelper.php
@@ -34,35 +34,39 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   Storyplayer/Cli
+ * @package   Storyplayer/Injectables
  * @author    Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright 2011-present Mediasift Ltd www.datasift.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
 
-namespace DataSift\Storyplayer\Cli;
+namespace DataSift\Storyplayer\Injectables;
 
 /**
- * support for Storyplayer's static config file manager
+ * Helper to find our hostname
  *
  * @category  Libraries
- * @package   Storyplayer/Cli
+ * @package   Storyplayer/Injectables
  * @author    Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright 2011-present Mediasift Ltd www.datasift.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
-trait Injectables_StaticConfigManagerSupport
+class HostnameHelper
 {
-	public $staticConfigManager;
+    static public function getHostname()
+    {
+        $hostname = getHostname();
 
-	public function initStaticConfigManagerSupport()
-	{
-		// create an object to manage the static config
-		$this->staticConfigManager = new StaticConfigManager;
+        // we get different results on different operating systems
+        // make sure the hostname is not the FQDN
+        $dotPos = strpos($hostname, '.');
+        if ($dotPos) {
+            $hostname = substr($hostname, 0, $dotPos);
+        }
 
-		// all done
-		return $this->staticConfigManager;
-	}
+        // all done
+        return $hostname;
+    }
 }
