+++ src/php/DataSift/Storyplayer/Injectables/TemplateEngineSupport.php
@@ -34,46 +34,47 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  * @category  Libraries
- * @package   Storyplayer/Cli
+ * @package   Storyplayer/Injectables
  * @author    Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright 2011-present Mediasift Ltd www.datasift.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
 
-namespace DataSift\Storyplayer\Cli;
+namespace DataSift\Storyplayer\Injectables;
+
+use Twig_Environment;
+use Twig_Loader_String;
 
 /**
- * support for working with the persistent, app-generated config
+ * support for rendering templates of any kind
  *
  * @category  Libraries
- * @package   Storyplayer/Cli
+ * @package   Storyplayer/Injectables
  * @author    Stuart Herbert <stuart.herbert@datasift.com>
  * @copyright 2011-present Mediasift Ltd www.datasift.com
  * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
  * @link      http://datasift.github.io/storyplayer
  */
-trait Injectables_RuntimeConfigSupport
+trait TemplateEngineSupport
 {
-	public $runtimeConfigManager;
-	public $runtimeConfig;
+	/**
+	 * an engine any part of Storyplayer can use to expand templates,
+	 * such as config files
+	 *
+	 * @var \Twig_Environment
+	 */
+	public $tempateEngine;
 
 	/**
 	 *
-	 * @return RuntimeConfigManager
+	 * @return void
 	 */
-	public function initRuntimeConfigSupport(Injectables $injectables)
+	public function initTemplateEngineSupport()
 	{
-		// create the runtime config's manager
-		$this->runtimeConfigManager = new RuntimeConfigManager();
-
-		// create the folder where we will store the persistent config
-		$this->runtimeConfigManager->makeConfigDir($injectables);
-
-		// load any config from the last time Storyplayer ran
-		$this->runtimeConfig = $this->runtimeConfigManager->loadRuntimeConfig();
-
-		// all done
-		return $this->runtimeConfigManager;
+        // we're going to expose this environment through Twig, to give
+        // us template support
+        $loader = new Twig_Loader_String();
+        $this->templateEngine   = new Twig_Environment($loader);
 	}
 }
