+++ src/php/DataSift/Storyplayer/PlayerLib/BasePlayer.php
@@ -43,7 +43,7 @@
 
 namespace DataSift\Storyplayer\PlayerLib;
 
-use DataSift\Storyplayer\Cli\Injectables;
+use DataSift\Storyplayer\Injectables;
 
 /**
  * our main base class for actually playing things
