+++ src/php/DataSift/Storyplayer/PlayerLib/Story/Player.php
@@ -43,7 +43,7 @@
 
 namespace DataSift\Storyplayer\PlayerLib;
 
-use DataSift\Storyplayer\Cli\Injectables;
+use DataSift\Storyplayer\Injectables;
 
 /**
  * the main class for animating a single story
